Move filter_data and to_snake_case to separate modules
